adding schematic pdf for people without KiCad
I also changed the scheamtic to reflect the known hardware issue where CLD1 was connected to PIN4. PIN4 is a straight up GND pin on the uC. So far this addresses the issues with the schematic. LAYOUT HAS NOT BEEN UPDATED YET. This is a @todo before doing another run of PCBs.